# docs/openapi.yaml
openapi: 3.1.0
info:
  title: discord-support-hub API
  version: 0.1.0
  summary: API-driven provisioning and governance of isolated private Discord support spaces.
  description: |
    The control surface (FR-11) for discord-support-hub, consumed by the backoffice
    (no human UI in v1). Postgres is the authorization source of truth; Discord is a
    projection the bot reconciles (three-layer truth model).

    Conventions:
      * **Mutating endpoints are asynchronous.** They validate, persist the *desired*
        state, enqueue a job, and return `202 Accepted` with a `job` handle. Poll
        `GET /jobs/{jobId}` for completion. (NFR-2: latency decoupled from Discord rate limits.)
      * **Read endpoints** are served from cache (TTL + write-invalidation) and return `200`.
      * **Authorization always resolves against Postgres**, never the Discord role (NFR-13).
      * **Idempotency:** mutating requests accept an `Idempotency-Key` header; a repeated key
        replays the original response instead of re-enqueueing (NFR-3).
      * **Role-per-merchant access.** Provisioning a space auto-creates a Discord role for the
        merchant and grants that role VIEW+SEND on the channel. A collaborator acquires access by
        joining through the merchant's native invite-with-role link (created once by the operator
        in the Discord client and stored via `PUT /merchants/{merchantId}/invite`), which the hub
...
  license:
    name: Apache-2.0
    identifier: Apache-2.0
...
servers:
...
security:
  - ServiceApiKey: []
...
paths:
...
  /channels/{id}/members:
    get:
      tags: [Spaces]
      operationId: listSpaceMembers
      summary: Users of a space with their role and merchant (FR-17).
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      responses:
        '200':
          description: Members of the space.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/SpaceMember' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /directory:
    get:
      tags: [Transversal]
      operationId: getDirectory
      summary: Global directory - spaces x users x role, bidirectional search (FR-18).
      parameters:
        - name: user_id
          in: query
          required: false
          description: Filter to spaces this user is in ("in what spaces is this user").
          schema: { type: string, format: uuid }
        - name: space_id
          in: query
          required: false
          description: Filter to users in this space ("who is in this space").
          schema: { type: string, format: uuid }
        - name: merchant_id
          in: query
          required: false
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Directory entries (space x user x role).
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DirectoryEntry' }
                  next_cursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /audit:
    get:
      tags: [Transversal]
      operationId: getAudit
      summary: Audit log of provisioning, membership, lifecycle, expulsion (FR-14).
      parameters:
        - name: merchant_id
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: space_id
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: action
          in: query
          required: false
          schema: { type: string }
        - name: since
          in: query
          required: false
          schema: { type: string, format: date-time }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Audit entries (newest first).
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AuditEntry' }
                  next_cursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ----------------------------------------------------------------------- #
  # JOBS (async polling)
  # ----------------------------------------------------------------------- #
...
components:
  securitySchemes:
    ServiceApiKey:
      type: http
      scheme: bearer
      description: |
        Opaque service API key issued to the backoffice (AuthZ Layer A). Sent as
        `Authorization: Bearer <key>`. Stored hashed at rest; instantly revocable. The future POC
        frontend uses a separate session principal (not this long-lived key) behind the same
        authorization layer.
  parameters:
...
    SpaceId:
      name: id
      in: path
      required: true
      description: Space (channel) id.
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid service API key.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    UserType:
      type: string
      enum: [agent, collaborator]
...
    SpaceMember:
      type: object
      required: [user_id, role]
      properties:
        user_id: { type: string, format: uuid }
        discord_user_id: { type: [string, 'null'] }
        display_name: { type: [string, 'null'] }
        merchant_id:
          type: [string, 'null']
          format: uuid
          description: The space's merchant (derived from the space, not a property of the collaborator). A collaborator is a global identity and may be a member of spaces across several merchants.
        email:
          type: [string, 'null']
          format: email
          description: The collaborator's work email (our label); the invite link is emailed here.
        role:
          type: string
          enum: [collaborator]
        invite_sent_at:
          type: [string, 'null']
          format: date-time
          description: When the merchant invite-with-role link was emailed to this collaborator (null = not yet sent).
        invited_by: { type: [string, 'null'], format: uuid }
        created_at: { type: string, format: date-time }
...
    DirectoryEntry:
      type: object
      description: One space x user x role row. The merchant is the space's merchant (derived from the space); for a collaborator it does not imply the user belongs to that merchant, since a collaborator may appear in spaces of several merchants.
      required: [space_id, user_id, role]
      properties:
        space_id: { type: string, format: uuid }
        space_name: { type: string }
        merchant_id:
          type: string
          format: uuid
          description: The space's merchant (derived from the space).
        merchant_name: { type: string }
        user_id: { type: string, format: uuid }
        user_display_name: { type: [string, 'null'] }
        role: { $ref: '#/components/schemas/UserType' }
    AuditEntry:
      type: object
      required: [id, action, created_at]
      properties:
        id: { type: integer, format: int64 }
        action:
          type: string
          description: e.g. space.provision, merchant.invite_set, collaborator.register, collaborator.invite_sent, collaborator.expel, agent.add, space.archive, reconcile.repair.
        actor_user_id: { type: [string, 'null'], format: uuid }
        merchant_id: { type: [string, 'null'], format: uuid }
        space_id: { type: [string, 'null'], format: uuid }
        target_user_id: { type: [string, 'null'], format: uuid }
        scope:
          type: [string, 'null']
          enum: [channel, server, null]
        detail:
          type: [object, 'null']
          additionalProperties: true
        created_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code, e.g. validation_error, forbidden, not_found, idempotency_conflict.
        message: { type: string }
        details:
          type: [object, 'null']
          additionalProperties: true